Create DTO

+++ primeiro-exemplo/src/main/java/com/github/primeiro_exemplo/services/ProdutoService.java
@@ -1,27 +1,43 @@
 package com.github.primeiro_exemplo.services;
 
+import com.github.primeiro_exemplo.PrimeiroExemploApplication;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.github.primeiro_exemplo.model.Produto;
+import com.github.primeiro_exemplo.model.exception.ResourceNotFoundException;
 import com.github.primeiro_exemplo.repository.ProdutoRepository;
+import com.github.primeiro_exemplo.shared.ProdutoDTO;
 
 @Service
 public class ProdutoService {
 
+    private final PrimeiroExemploApplication primeiroExemploApplication;
     @Autowired
     private ProdutoRepository produtoRepository;
+
+    ProdutoService(PrimeiroExemploApplication primeiroExemploApplication) {
+        this.primeiroExemploApplication = primeiroExemploApplication;
+    }
 
     /**
      * Método para Retorna uma lista de Produtos.
      * 
      * @return Lista de produtos.
      */
-    public List<Produto> obterTodos() {
-        return produtoRepository.findAll();
+    public List<ProdutoDTO> obterTodos() {
+
+        List<Produto> produtos = produtoRepository.findAll();
+
+        return produtos.stream()
+        .map(produto -> new ModelMapper()
+        .map(produto, ProdutoDTO.class))
+        .collect(Collectors.toList());
     }
 
     /**
@@ -30,8 +46,19 @@ public class ProdutoService {
      * @param id do produto que será localizado.
      * @return Retona um produto caso seja endontrado.
      */
-    public Optional<Produto> obterPorId(Integer id) {
-        return produtoRepository.findById(id);
+    public Optional<ProdutoDTO> obterPorId(Integer id) {
+        // Obtendo optional de produto pelo id.
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        // Se não encontrar, lança exception
+        if(produto.isEmpty()){
+            throw new ResourceNotFoundException("produto com Id:"+id+"não encontrado");
+        }
+        // Convertendo o meu optional de produto em um produtoDTO
+        ProdutoDTO dto = new ModelMapper().map(produto.get(), ProdutoDTO.class);
+
+        // Criando e retornando um optional de produtoDTO
+        return Optional.of(dto);
     }
 
     /**
@@ -40,9 +67,23 @@ public class ProdutoService {
      * @param produto que será adicionado .
      * @return o produto que foi adicionado na lista.
      */
-    public Produto adicionar(Produto produto) {
-        // Poderia ter uma regra de negocios aqui para validar o produto;
-        return produtoRepository.saveAndFlush(produto);
+    public ProdutoDTO adicionar(ProdutoDTO produtoDTO) {
+        // Removendo id para conseguir fazer o cadastro
+       produtoDTO.setId(null);
+
+       // Criar um objeto de mapeamento.
+       ModelMapper mapper = new ModelMapper();
+
+       // Converter o nosso ProdutoDTO em um Produto. 
+       Produto produto = mapper.map(produtoDTO, Produto.class);
+
+       // Salvar o produto do banco.
+       produto = produtoRepository.save(produto);
+
+       produto.setId(produto.getId());
+
+       // Retornando um produtoDTO atualizado.
+        return produtoDTO;
     }
 
     /**
@@ -51,7 +92,14 @@ public class ProdutoService {
      * @param id do produto a ser deletado.
      */
     public void deletar(Integer id) {
-        // Aqui poderia ter uma logica
+        // Verificar se o produto exite
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        // Se não existir lança uma exception
+        if(produto.isEmpty()){
+            throw new ResourceNotFoundException("Não foi possivel deletar o produto com o id:"+id+ " produto não encontrado");
+        }
+        // Deleta o produto pelo id.
         produtoRepository.deleteById(id);
     }
 
@@ -61,10 +109,22 @@ public class ProdutoService {
      * @param produto que será atualizado.
      * @return Retorna o produto após atualizar a lista.
      */
-    public Produto atualizar(Integer id, Produto produto) {
-        // Ter uma validação do Id
-        produto.setId(id);
-        return produtoRepository.saveAndFlush(produto);
+    public ProdutoDTO atualizar(Integer id, ProdutoDTO produtoDTO) {
+       // Passar o Id para o ProdutoDTO
+        produtoDTO.setId(id);
+
+       // Criar um objeto de mapeamento.
+        ModelMapper mapper = new ModelMapper();
+
+       // Converter um produtoDTO em um Produto.
+       Produto produto = mapper.map(produtoDTO, Produto.class);
+
+       // Atualiar o produto no Banco de  dados.
+        produtoRepository.save(produto);
+        
+        // Retorna o produtoDTO atualizado
+        return produtoDTO;
+
     }
 
 }
